+++ src/main/java/com/deensoft/mathfun/printer/FizzPrinter.java
@@ -10,7 +10,7 @@ public class FizzPrinter implements FunPrinter {
     private static Logger LOGGER = LoggerFactory.getLogger(FizzPrinter.class);
 
     @Override
-    public void print(short num) {
+    public void print(int num) {
         LOGGER.info(FIZZ);
     }
 }
