+++ src/main/java/com/deensoft/mathfun/printer/FunPrinter.java
@@ -1,5 +1,5 @@
 package com.deensoft.mathfun.printer;
 
 public interface FunPrinter {
-    void print(short num);
+    void print(int num);
 }
